refactor: split file for vpm_manifest
